refactor(parser): relocate API-key and node-list parsing
Moves parseNodeIdMap and the key/node merge (renamed mergeApiKeysWithNodes,
now taking the data string it actually used) out of NodeCredsTable into
Parser, alongside the other broker XML parsers. NodeCredsTable consumes them
and routes its status errors through StatusToast.
